something update

+++ app/controllers/nextbbs/boards_controller.rb
@@ -3,11 +3,13 @@ require_dependency "nextbbs/application_controller"
 module Nextbbs
   class BoardsController < ApplicationController
     before_action :set_board, only: [:show, :edit, :update, :destroy, :control]
-    before_action :_authenticate_with, only: [:new, :create, :update, :edit, :destroy, :control]
+    before_action :_authenticate_with, only: [:new, :create, :update, :edit, :destroy, :control, :yours]
+    before_action :validate_user, only: [:control, :edit, :update, :destroy]
 
     # GET /boards
     def index
-      @boards = Board.all
+      # @boards = Board.all
+      @boards = Board.published
     end
 
     # GET /boards/1
@@ -15,7 +17,10 @@ module Nextbbs
       @topics = @board.topics
       respond_to do |format|
         case @board.status.to_sym
-        when :deleted, :unpublished then
+        when :deleted
+          format.html { head :not_found }
+          format.json { head :not_found } # これでいいかな？
+        when :unpublished
           if _current_user == @board.owner
             format.html { render }
             format.json { render }
@@ -24,11 +29,17 @@ module Nextbbs
             format.html { head :not_found }
             format.json { head :not_found } # これでいいかな？
           end
-        when :published then
+        when :published
           format.html { render }
           format.json { render }
         end
       end
+    end
+
+    # GET /boards/yours
+    def yours
+      # TODO: Board.find でUser.idを絞ったほうが良いかもね
+      @boards = _current_user.nextbbs_boards
     end
 
     # GET /boards/new
@@ -38,6 +49,7 @@ module Nextbbs
 
     # GET /boards/1/edit
     def edit
+      @topics = @board.topics
     end
 
     # GET /boards/1/control
@@ -50,7 +62,10 @@ module Nextbbs
       @board = Board.new(board_params)
       @board.owner = _current_user
       if @board.save
-        redirect_to @board, notice: "Board was successfully created."
+        respond_to do |format|
+          format.html { redirect_to yours_boards_path, notice: t(".create_success") }
+          format.json { head :no_content }
+        end
       else
         logger.debug @board.errors.inspect
         render :new
@@ -59,8 +74,9 @@ module Nextbbs
 
     # PATCH/PUT /boards/1
     def update
+      # Ownerチェック
       if @board.update(board_params)
-        redirect_to @board, notice: "Board was successfully updated."
+        redirect_to @board, notice: t(".update_success")
       else
         render :edit
       end
@@ -68,13 +84,22 @@ module Nextbbs
 
     # DELETE /boards/1
     def destroy
-      respond_to do |format|
-        if @board.destroy
-          format.html { redirect_to boards_url, notice: "Board was successfully destroyed." }
-          format.json { head :no_content }
-        else
-          format.html { redirect_to boards_url, notice: "Board was not destroyed. Tell Server Administrator" }
-          format.json { head :no_content }
+      logger.debug destroy_board_params
+      unless destroy_board_params[:destroy_seed] && destroy_board_params[:destroy_seed] == @board.hash_token
+        flash[alert] = "hash_tokenが違います。"
+        redirect_to edit_board_path(@board)
+      else
+        logger.debug "掲示板を削除します"
+        respond_to do |format|
+          if @board.destroy
+            format.html { redirect_to yours_boards_url, notice: t(".destroy_success") }
+            format.js { render ajax_redirect_to(yours_boards_url) }
+            format.json { head :no_content }
+          else
+            format.html { redirect_to boards_url, notice: t(".destroy_failed") }
+            format.js { render ajax_redirect_to(boards_url) }
+            format.json { head :no_content }
+          end
         end
       end
     end
@@ -88,7 +113,18 @@ module Nextbbs
 
     # Only allow a trusted parameter "white list" through.
     def board_params
-      params.require(:board).permit(:title, :description, :name, :status, :owner_id)
+      # params.require(:form_board).permit(:title, :description, :name, :status)
+      params.require(:board).permit(:title, :description, :name, :status)
+    end
+
+    def destroy_board_params
+      params.require(:board).permit(:destroy_seed)
+    end
+
+    def validate_user
+      if @board.owner != _current_user
+        redirect_to board_path(@board), alert: t(".validate_failed")
+      end
     end
   end
 end
